moved NavMenu to appropriate folder
